net: 6lowpan: Fix 6lowpan fragmentation
IP packet passing through 6lowpan compression and 802.15.4
fragmentation has an issue. (RFC 6282 and 4944) RFC4944, 5.3
says "fragment header's datagram_size and datagram_offset
values as the size and offset of the IPv6 datagram before
compression".

At the moment datagram size and offset values are after
compression. Due to multiple bearer (15.4 and bluetooth) support
and multiple incoming buffers support functionality is divided
to compression and fragmentation parts. And datagram size and
offset values for BT are after compression. Small packet (which
doesn't need fragmentation) doesn't go through this. So cacheing
compressed and uncompressed header lengths and with hdr difference
preparing fragments as per original buffer (total buf size and offset).

Jira: ZEP-208

+++ include/net/l2_buf.h
@@ -64,7 +64,7 @@ struct l2_buf {
 	/* 6LoWPAN pointers */
 	uint8_t *packetbuf_ptr;
 	uint8_t packetbuf_hdr_len;
-	int packetbuf_payload_len;
+	uint8_t packetbuf_payload_len;
 	uint8_t uncomp_hdr_len;
 	int last_tx_status;
 #if defined(CONFIG_NETWORKING_WITH_15_4)
